--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/dto/ExportAllPointHistoryResponse.kt
@@ -1,6 +1,0 @@
-package team.aliens.dms.domain.point.dto
-
-class ExportAllPointHistoryResponse(
-    val fileName: String,
-    val file: ByteArray
-)
