--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/dto/QueryPointHistoryResponse.kt
@@ -1,6 +1,0 @@
-package team.aliens.dms.domain.point.dto
-
-data class QueryPointHistoryResponse(
-    val totalPoint: Int,
-    val points: List<PointHistoryDto>
-)
