--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/dto/QueryStudentPointHistoryResponse.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.point.dto
-
-data class QueryStudentPointHistoryResponse(
-    val pointHistories: List<PointHistoryDto>
-)
